Carve out the previously split up methods.
This commit also removes the last uses of underscore.
