+++ lib/unexpectedExpress.js
@@ -1,12 +1,11 @@
 const http = require('http');
-const stream = require('stream');
-const isStream = require('is-stream');
-const _ = require('underscore');
-const FormData = require('form-data');
-const qs = require('qs');
 const messy = require('messy');
+const qs = require('qs');
+const stream = require('stream');
 
+const classifyRequestBodyAndUpdateHttpRequest = require('./classifyRequestBodyAndUpdateHttpRequest');
 const resolveExpectedResponseProperties = require('./resolveExpectedResponseProperties');
+const resolveRequestBody = require('./resolveRequestBody');
 
 function applyRequestBodyToToReq(
   req,
@@ -50,106 +49,6 @@ function applyRequestBodyToToReq(
   } else {
     req.push(null);
   }
-}
-
-function classifyRequestBodyAndUpdateHttpRequest(requestBody, httpRequest) {
-  let hasFormData = false;
-  let hasStream = false;
-
-  if (typeof requestBody === 'undefined') {
-    return null;
-  }
-
-  httpRequest.headers.set('Transfer-Encoding', 'chunked');
-  if (requestBody.pipe) {
-    hasStream = true;
-    if (
-      requestBody.constructor &&
-      requestBody.constructor.name === 'FormData'
-    ) {
-      hasFormData = true;
-      if (!httpRequest.headers.has('Content-Type')) {
-        httpRequest.headers.set(
-          'Content-Type',
-          `multipart/form-data; boundary=${requestBody.getBoundary()}`
-        );
-      }
-    }
-  } else {
-    if (typeof requestBody === 'object' && !Buffer.isBuffer(requestBody)) {
-      if (!httpRequest.headers.has('Content-Type')) {
-        httpRequest.headers.set('Content-Type', 'application/json');
-      }
-    }
-
-    if (
-      !httpRequest.headers.has('Content-Length') &&
-      !httpRequest.headers.has('Transfer-Encoding')
-    ) {
-      httpRequest.headers.set('Content-Length', String(requestBody.length));
-    }
-  }
-
-  return {
-    hasFormData,
-    hasStream
-  };
-}
-
-function resolveRequestBody(requestProperties, httpRequest) {
-  let requestBody = requestProperties.body;
-
-  if (requestProperties.form) {
-    const requestForm = requestProperties.form;
-    if (typeof requestForm !== 'string') {
-      requestBody = qs.stringify(requestForm);
-    } else {
-      requestBody = requestForm;
-    }
-
-    if (!httpRequest.headers.has('Content-Type')) {
-      httpRequest.headers.set(
-        'Content-Type',
-        'application/x-www-form-urlencoded'
-      );
-    }
-  } else if (requestProperties.formData) {
-    if (requestBody) {
-      throw new Error(
-        'unexpected-express: The "body" and "formData" options are not supported together'
-      );
-    }
-
-    requestBody = new FormData();
-
-    Object.keys(requestProperties.formData).forEach(name => {
-      let value = requestProperties.formData[name];
-
-      let options = {};
-
-      if (isStream.readable(value) && value.path) {
-        options.filename = value.path;
-      } else if (typeof value === 'object' && !Buffer.isBuffer(value)) {
-        options = _.extend({}, value);
-        value = options.value;
-        delete options.value;
-        if (options.fileName) {
-          options.filename = options.fileName;
-          delete options.fileName;
-        }
-      }
-      requestBody.append(name, value, options);
-    });
-
-    delete requestProperties.formData;
-  } else if (
-    'unchunkedBody' in requestProperties ||
-    'rawBody' in requestProperties
-  ) {
-    requestBody = httpRequest.body;
-  }
-
-  return requestBody;
 }
 
 module.exports = {
@@ -205,7 +104,7 @@ module.exports = {
         const requestProperties =
           typeof value.request === 'string'
             ? { url: value.request }
-            : _.extend({}, value.request);
+            : { ...value.request };
 
         const httpRequest = new messy.HttpRequest({
           method: requestProperties.method,
@@ -322,7 +221,7 @@ module.exports = {
         });
         req.method = httpRequest.method;
         req.url = httpRequest.requestLine.url;
-        _.extend(req, requestProperties);
+        Object.assign(req, requestProperties);
 
         applyRequestBodyToToReq(
           req,
@@ -334,7 +233,7 @@ module.exports = {
         );
 
         const res = new http.ServerResponse(req);
-        _.extend(res, requestProperties.res); // Allows for specifying eg. res.locals
+        Object.assign(res, requestProperties.res); // Allows for specifying eg. res.locals
         delete requestProperties.res;
         res.locals = res.locals || {};
 
@@ -416,7 +315,7 @@ module.exports = {
             }
           })
           .then(() => {
-            _.extend(context, {
+            Object.assign(context, {
               req,
               res,
               httpRequest,
