+++ package.json
@@ -63,7 +63,6 @@
     "is-stream": "^2.0.0",
     "messy": "^6.6.1",
     "qs": "^6.5.1",
-    "underscore": "^1.6.0",
     "unexpected-messy": "^9.0.0"
   },
   "nyc": {
